+++ src/components/spec-document/HttpOperation.vue
@@ -193,7 +193,7 @@ const securitySchemeGroupList = computed<Array<SecuritySchemeGroup>>(() => {
       let title = ''
       let key = ''
       secGroup.forEach((scheme) => {
-        title = title.length ? title + ` & ${scheme.key}` : scheme.key,
+        title = title.length ? title + ` & ${scheme.key}` : scheme.key
         key = key.length ? key + `-${scheme.key.replace(' ', '-')}` : scheme.key.replace(' ', '-')
       })
       schemeGroupList.push({
@@ -346,7 +346,7 @@ watch(() => ({ id: props.data.id, excludeNotRequired: excludeNotRequired.value }
     margin-bottom: var(--kui-space-90, $kui-space-90);
   }
 
-  .http-operation-container  {
+  .http-operation-container {
     display: grid;
     gap: var(--kui-space-90, $kui-space-90);
     grid-template-columns: auto clamp($spec-renderer-secondary-column-min-width, 40%, $spec-renderer-secondary-column-max-width);