+++ src/types/security.ts
@@ -1,7 +1,7 @@
 import type { HttpSecurityScheme } from '@stoplight/types'
 
 export interface SecuritySchemeGroup {
-  title: string;
-  key: string;
-  schemeList: HttpSecurityScheme[];
+  title: string
+  key: string
+  schemeList: HttpSecurityScheme[]
 }